requirements.txt
azure-core==1.17.0
azure-storage-blob==12.8.1
black==21.8b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
idna==3.2
isodate==0.6.0
msrest==0.6.21
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.2
pathspec==0.9.0
platformdirs==2.3.0
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
